Move cache to top-level module
