+++ crates/cache/src/lib.rs
@@ -1,3 +1,4 @@
+use anyhow::Result;
 use base64::Engine;
 use log::{debug, trace, warn};
 use serde::{Deserialize, Serialize};
@@ -6,14 +7,126 @@ use std::hash::Hash;
 use std::hash::Hasher;
 use std::io::Write;
 use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicUsize, Ordering::SeqCst};
+use std::sync::Arc;
+use std::time::Duration;
 use std::{fs, io};
 
 #[macro_use] // for tests
 mod config;
 mod worker;
 
-pub use config::{create_new_config, Cache, CacheConfig};
+pub use config::{create_new_config, CacheConfig};
 use worker::Worker;
+
+/// Global configuration for how the cache is managed
+#[derive(Debug, Clone)]
+pub struct Cache {
+    config: CacheConfig,
+    worker: Worker,
+    state: Arc<CacheState>,
+}
+
+macro_rules! generate_config_setting_getter {
+    ($setting:ident: $setting_type:ty) => {
+        /// Returns `$setting`.
+        ///
+        /// Panics if the cache is disabled.
+        pub fn $setting(&self) -> $setting_type {
+            self.config.$setting()
+        }
+    };
+}
+
+impl Cache {
+    /// Builds a [`Cache`] from the configuration and spawns the cache worker.
+    ///
+    /// If you want to load the cache configuration from a file, use [`CacheConfig::from_file`].
+    /// You can call [`CacheConfig::new`] for the default configuration.
+    ///
+    /// # Errors
+    /// Returns an error if the configuration is invalid.
+    pub fn new(mut config: CacheConfig) -> Result<Self> {
+        config.validate()?;
+        Ok(Self {
+            worker: Worker::start_new(&config),
+            config,
+            state: Default::default(),
+        })
+    }
+
+    /// Loads cache configuration specified at `path`.
+    ///
+    /// This method will read the file specified by `path` on the filesystem and
+    /// attempt to load cache configuration from it. This method can also fail
+    /// due to I/O errors, misconfiguration, syntax errors, etc. For expected
+    /// syntax in the configuration file see the [documentation online][docs].
+    ///
+    /// Passing in `None` loads cache configuration from the system default path.
+    /// This is located, for example, on Unix at `$HOME/.config/wasmtime/config.toml`
+    /// and is typically created with the `wasmtime config new` command.
+    ///
+    /// # Errors
+    ///
+    /// This method can fail due to any error that happens when loading the file
+    /// pointed to by `path` and attempting to load the cache configuration.
+    ///
+    /// [docs]: https://bytecodealliance.github.io/wasmtime/cli-cache.html
+    pub fn from_file(path: Option<&Path>) -> Result<Self> {
+        let config = CacheConfig::from_file(path)?;
+        Self::new(config)
+    }
+
+    generate_config_setting_getter!(worker_event_queue_size: u64);
+    generate_config_setting_getter!(baseline_compression_level: i32);
+    generate_config_setting_getter!(optimized_compression_level: i32);
+    generate_config_setting_getter!(optimized_compression_usage_counter_threshold: u64);
+    generate_config_setting_getter!(cleanup_interval: Duration);
+    generate_config_setting_getter!(optimizing_compression_task_timeout: Duration);
+    generate_config_setting_getter!(allowed_clock_drift_for_files_from_future: Duration);
+    generate_config_setting_getter!(file_count_soft_limit: u64);
+    generate_config_setting_getter!(files_total_size_soft_limit: u64);
+    generate_config_setting_getter!(file_count_limit_percent_if_deleting: u8);
+    generate_config_setting_getter!(files_total_size_limit_percent_if_deleting: u8);
+
+    /// Returns path to the cache directory.
+    ///
+    /// Panics if the cache directory is not set.
+    pub fn directory(&self) -> &PathBuf {
+        &self.config.directory()
+    }
+
+    #[cfg(test)]
+    fn worker(&self) -> &Worker {
+        &self.worker
+    }
+
+    /// Returns the number of cache hits seen so far
+    pub fn cache_hits(&self) -> usize {
+        self.state.hits.load(SeqCst)
+    }
+
+    /// Returns the number of cache misses seen so far
+    pub fn cache_misses(&self) -> usize {
+        self.state.misses.load(SeqCst)
+    }
+
+    pub(crate) fn on_cache_get_async(&self, path: impl AsRef<Path>) {
+        self.state.hits.fetch_add(1, SeqCst);
+        self.worker.on_cache_get_async(path)
+    }
+
+    pub(crate) fn on_cache_update_async(&self, path: impl AsRef<Path>) {
+        self.state.misses.fetch_add(1, SeqCst);
+        self.worker.on_cache_update_async(path)
+    }
+}
+
+#[derive(Default, Debug)]
+struct CacheState {
+    hits: AtomicUsize,
+    misses: AtomicUsize,
+}
 
 /// Module level cache entry.
 pub struct ModuleCacheEntry<'cache>(Option<ModuleCacheEntryInner<'cache>>);
